+++ plugins/api-docs/src/components/ApiDefinitionCard/index.ts
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-export type { ApiDefinitionWidget } from './ApiDefinitionCard';
-export {
-  ApiDefinitionCard,
-  defaultDefinitionWidgets,
-} from './ApiDefinitionCard';
+export { ApiDefinitionCard } from './ApiDefinitionCard';
+export { defaultDefinitionWidgets } from './ApiDefinitionWidget';
+export type { ApiDefinitionWidget } from './ApiDefinitionWidget';
+export { ApiTypeTitle } from './ApiTypeTitle';